First pass
These changes split the current reference section, adding a "Getting Started" section that provides simplicity to what is necessary information for most users. The Reference section is now reserved for material to be called only when needed.

This first pass does not include any changes to the Feature Guides or newly retitled "Technical Data" section.

"Getting Started" button on homepage is broken in this version. This is temporary until committing.